+++ src/views/Tags.vue
@@ -1,5 +1,7 @@
 <template>
-  <div class="wrapper">
+  <div
+    class="wrapper"
+    :class="{ 'has-scroll-to-top-bar': showScrollToTopBar }">
 
     <vue-headful :title="title | formatToConfiguredLetterCase" />
 
@@ -10,7 +12,7 @@
         :lastQuery="lastQuery"
         :wantsAdditionalTags="wantsAdditionalTags" />
 
-      <hr>
+      <hr id="results-separator">
 
       <div class="table-container" v-if="formattedTags.length">
 
@@ -60,6 +62,8 @@
 
     </section>
 
+    <scroll-to-top-bar v-if="showScrollToTopBar" />
+
   </div>
 </template>
 
@@ -72,8 +76,10 @@ import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
 import config from '@/config'
 import queryFormatter from '@/util/query-formatter'
 import tagsHelper from '@/util/tags-helper'
+import visibilityHelper from '@/util/visibility-helper'
 
 import Search from '@/components/tags/Search'
+import ScrollToTopBar from '@/components/general/ScrollToTopBar'
 
 let scrollListener
 
@@ -82,6 +88,7 @@ export default {
   data: function () {
     return {
       wantsAdditionalTags: false,
+      showScrollToTopBar: false,
       title: `Tags – ${config.title}`
     }
   },
@@ -110,19 +117,16 @@ export default {
       colors: state => state.settings.colors
     })
   },
-  methods: {
-    bottomIsVisible: function () {
-      const scrollY = window.scrollY
-      const isVisible = document.documentElement.clientHeight
-      const pageHeight = document.documentElement.scrollHeight
-      const isAtBottomOfPage = (isVisible + scrollY) >= (pageHeight - 48)
+  mounted: function () {
+    this.showScrollToTopBar = visibilityHelper.isPageScrolledPastElement(
+      'results-separator'
+    )
 
-      return isAtBottomOfPage || ((pageHeight - 48) < isVisible)
-    }
-  },
-  created: function () {
     scrollListener = throttle(() => {
-      this.wantsAdditionalTags = this.bottomIsVisible()
+      this.wantsAdditionalTags = visibilityHelper.isBottomOfPageVisible(88)
+      this.showScrollToTopBar = visibilityHelper.isPageScrolledPastElement(
+        'results-separator'
+      )
     }, 50)
 
     window.addEventListener(
@@ -148,7 +152,8 @@ export default {
   },
   components: {
     FontAwesomeIcon,
-    Search
+    Search,
+    ScrollToTopBar
   }
 }
 </script>