+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
++ Added a scroll to top bar to the files and tags views
+
 ## [3.0.2] - 2018-12-31
 
 ### Fixed